scaffolder-backend: move actions to separate actions dir
