+++ plugins/scaffolder-backend/src/scaffolder/actions/TemplateActionRegistry.ts
